made changes to the controllers, models and migrations

+++ Controllers/ProjectsController.cs
@@ -1,8 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using GymLinkPro.Data;
@@ -33,23 +32,55 @@ namespace GymLinkPro.Controllers
         public async Task<ActionResult<Project>> GetProject(int id)
         {
             var project = await _context.Projects.FindAsync(id);
-
             if (project == null)
             {
                 return NotFound();
             }
-
             return project;
         }
 
+        // POST: api/Projects
+        [HttpPost]
+        public async Task<ActionResult<Project>> PostProject(Project project)
+        {
+            int currentUserId = GetCurrentUserId();
+            project.CreatorId = currentUserId;
+
+            _context.Projects.Add(project);
+            await _context.SaveChangesAsync();
+
+            // Add creator as ProjectMember with Admin role
+            var adminMember = new ProjectMember
+            {
+                ProjectId = project.ProjectId,
+                MemberId = currentUserId,
+                Role = "Admin"
+            };
+            _context.ProjectMembers.Add(adminMember);
+            await _context.SaveChangesAsync();
+
+            return CreatedAtAction(nameof(GetProject), new { id = project.ProjectId }, project);
+        }
+
         // PUT: api/Projects/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
         public async Task<IActionResult> PutProject(int id, Project project)
         {
             if (id != project.ProjectId)
             {
                 return BadRequest();
+            }
+
+            int currentUserId = GetCurrentUserId();
+            var membership = await _context.ProjectMembers
+                .FirstOrDefaultAsync(pm => pm.ProjectId == id && pm.MemberId == currentUserId);
+
+            bool isAdminOrCoAdmin = membership != null &&
+                (membership.Role == "Admin" || membership.Role == "Co-Admin");
+
+            if (!isAdminOrCoAdmin)
+            {
+                return Forbid();
             }
 
             _context.Entry(project).State = EntityState.Modified;
@@ -73,17 +104,6 @@ namespace GymLinkPro.Controllers
             return NoContent();
         }
 
-        // POST: api/Projects
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPost]
-        public async Task<ActionResult<Project>> PostProject(Project project)
-        {
-            _context.Projects.Add(project);
-            await _context.SaveChangesAsync();
-
-            return CreatedAtAction("GetProject", new { id = project.ProjectId }, project);
-        }
-
         // DELETE: api/Projects/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteProject(int id)
@@ -92,6 +112,18 @@ namespace GymLinkPro.Controllers
             if (project == null)
             {
                 return NotFound();
+            }
+
+            int currentUserId = GetCurrentUserId();
+            var membership = await _context.ProjectMembers
+                .FirstOrDefaultAsync(pm => pm.ProjectId == id && pm.MemberId == currentUserId);
+
+            bool isAdminOrCoAdmin = membership != null &&
+                (membership.Role == "Admin" || membership.Role == "Co-Admin");
+
+            if (!isAdminOrCoAdmin)
+            {
+                return Forbid();
             }
 
             _context.Projects.Remove(project);
@@ -103,6 +135,16 @@ namespace GymLinkPro.Controllers
         private bool ProjectExists(int id)
         {
             return _context.Projects.Any(e => e.ProjectId == id);
+        }
+
+        private int GetCurrentUserId()
+        {
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userIdClaim != null && int.TryParse(userIdClaim.Value, out int userId))
+            {
+                return userId;
+            }
+            throw new UnauthorizedAccessException("User ID not found in claims.");
         }
     }
 }
